Rename config models. Fix tim incl
